added user_query to the scenarios
