+++ tests/blackbox/data/evaluation/namespace-scoped/kyma_app_apirule_broken/scenario.yml
@@ -1,5 +1,6 @@
 id: kyma-app-apirule-broken
 description: The Application uses an API Rule that has two accessStrategies handler, which will lead to an error stead
+user_query: Why is the APIRule status showing ERROR?
 resource:
   kind: APIRule
   api_version: gateway.kyma-project.io/v1beta1
